requirements.txt
python-dotenv>=1.0.1
dj-database-url>=2.2
django-allauth>=65.0.0
stripe>=11.2.0
whitenoise>=6.7.0
gunicorn>=22.0
psycopg2-binary>=2.9
stripe>=11.2.0
Pillow>=11.0.0
python-dotenv>=1.0.1
